+++ extensions/iceworks-material-helper/src/extension.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import { connectService, getHtmlForWebview } from '@iceworks/vscode-webview/lib/vscode';
 import { initExtension, registerCommand } from '@iceworks/common-service';
 import { autoSetContext as autoSetContextByProject } from '@iceworks/project-service';
+import { ICEWORKS_ICON_PATH } from '@iceworks/constant';
 import services from './services/index';
 import propsAutoComplete from './propsAutoComplete';
 import autoFillContent from './autoFillContent';
@@ -50,6 +51,7 @@ export function activate(context: vscode.ExtensionContext) {
         },
       );
       materialImporterWebviewPanel.webview.html = getHtmlForWebview(extensionPath, 'materialimporter');
+      materialImporterWebviewPanel.iconPath = vscode.Uri.parse(ICEWORKS_ICON_PATH);
       materialImporterWebviewPanel.onDidDispose(
         () => {
           materialImporterWebviewPanel = undefined;
@@ -85,6 +87,7 @@ export function activate(context: vscode.ExtensionContext) {
       },
     );
     webviewPanel.webview.html = getHtmlForWebview(extensionPath, 'componentcreator');
+    webviewPanel.iconPath = vscode.Uri.parse(ICEWORKS_ICON_PATH);
     connectService(webviewPanel, context, { services, recorder });
   }
   subscriptions.push(
@@ -104,6 +107,7 @@ export function activate(context: vscode.ExtensionContext) {
       },
     );
     webviewPanel.webview.html = getHtmlForWebview(extensionPath, 'pagegenerator');
+    webviewPanel.iconPath = vscode.Uri.parse(ICEWORKS_ICON_PATH);
     connectService(webviewPanel, context, { services, recorder });
   }
   subscriptions.push(
@@ -123,6 +127,7 @@ export function activate(context: vscode.ExtensionContext) {
       },
     );
     webviewPanel.webview.html = getHtmlForWebview(extensionPath, 'pagecreator');
+    webviewPanel.iconPath = vscode.Uri.parse(ICEWORKS_ICON_PATH);
     connectService(webviewPanel, context, { services, recorder });
   }
   subscriptions.push(